Refactor top level pages into directories for nested routing
