+++ react-app/src/pages/News/index.js
